tutors/accounts/loginView.py
from django.shortcuts import render, redirect
from django.http import HttpResponse
from django.contrib.auth import login, authenticate, logout

from account.forms import RegistrationForm, AccountAuthenticationForm


def logout_view(request):
	logout(request)
	return redirect("home")


def login_view(request, *args, **kwargs):
	context = {}

	user = request.user
	if user.is_authenticated: 
		return redirect("home")

	destination = get_redirect_if_exists(request)
	print("destination: " + str(destination))

	if request.POST:
		form = AccountAuthenticationForm(request.POST)

        # Can also place this in the save() in the forms.py 
		if form.is_valid():
			email = request.POST['email']
			password = request.POST['password']
			user = authenticate(email=email, password=password)

			if user:
				login(request, user)
				if destination:
					return redirect(destination)
				return redirect("home")

	else:
		form = AccountAuthenticationForm()

	context['login_form'] = form

	return render(request, "account/login.html", context)




def edit_account_view(request, *args, **kwargs):
	if not request.user.is_authenticated:
		return redirect("login")
	user_id = kwargs.get("user_id")
	account = Account.objects.get(pk=user_id)

    #this makes sure that user is editing his own account 
    # not editing sb else's account 
	if account.pk != request.user.pk:
		return HttpResponse("You cannot edit someone elses profile.")
	context = {}

    # request.files for including an image or files 
    #instnace = request.user is the account we are submitting to the AccountUpdateForm()
    # basically it's the model Account 
	if request.method = ""post":
		form = AccountUpdateForm(request.POST, request.FILES, instance=request.user)

		if form.is_valid():
			form.save()
			new_username = form.cleaned_data['username']

            # account: view, name ="view" in the url.py 
			return redirect("account:view", user_id=account.pk)
		else:

			form = AccountUpdateForm(request.POST, instance=request.user,
       
            #the initival value set to the form 
            	initial={
					"id": account.pk,
					"email": account.email, 
					"username": account.username,
					"profile_image": account.profile_image,
					"hide_email": account.hide_email,
				}
			)
			context['form'] = form
	
    # This is for the 'get' request 
    else:
		form = AccountUpdateForm(
			initial={
					"id": account.pk,
					"email": account.email, 
					"username": account.username,
					"profile_image": account.profile_image,
					"hide_email": account.hide_email,
				}
			)
		context['form'] = form

    #Set a maxmium on the data size 
	context['DATA_UPLOAD_MAX_MEMORY_SIZE'] = settings.DATA_UPLOAD_MAX_MEMORY_SIZE
	# return render(request, "account/edit_account.html", context)
	return Response(context)



def get_redirect_if_exists(request):
	redirect = None
	if request.GET:
		if request.GET.get("next"):
			redirect = str(request.GET.get("next"))
	return redirect


