tutors/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
fire==0.1.2
Flask==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
numpy==1.20.3
opencv-python==4.5.2.52
peewee==2.8.8
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
rope==0.19.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
termcolor==1.1.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
virtualenv==20.4.6
Werkzeug==2.0.1
Workon==0.8.2
zope.interface==5.4.0